requirements.txt
jishaku
toml~=0.10.2
uvloop~=0.14.0
parsedatetime~=2.6
pytz~=2021.1
discord.py[voice]>=1.5.0
tortoise-orm[accel]
aerich
asyncpg
python-rapidjson
ciso8601
Babel~=2.9.1
discord.py-stubs
python-dateutil~=2.8.2
flask
aiohttp_cors
aiohttp~=3.7.4.post0
profanity